IbanCccBank/requirements.txt
Django==2.1.8
psycopg2-binary==2.8.2
django-bootstrap-modal-forms==1.4.1
django-widget-tweaks==1.4.2
social-auth-app-django==3.1.0
schwifty==2018.9.1

